cleaned up app.js
